+++ tests/database/factories/EntityWithExpiredAtFactory.php
@@ -1,0 +1,17 @@
+<?php
+
+/*
+ * This file is part of Laravel Eloquent Flag.
+ *
+ * (c) Anton Komarev <a.komarev@cybercog.su>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+$factory->define(\Cog\Flag\Tests\Stubs\Models\Inverse\EntityWithExpiredAt::class, function (\Faker\Generator $faker) {
+    return [
+        'name' => $faker->word,
+        'expired_at' => null,
+    ];
+});